Move camera endpoints to an independent file.
