+++ run_processor_api.py
@@ -1,25 +1,26 @@
 #!/usr/bin/python3
 import argparse
-from multiprocessing import Process
-import threading
+from api.settings import Settings
 from libs.config_engine import ConfigEngine
+
 import logging
 
 logger = logging.getLogger(__name__)
 
+
 def start_api(config):
     from api.processor_api import ProcessorAPI
-    api = ProcessorAPI(config)
-    
+    api = ProcessorAPI()
     logger.info("API Started.")
     api.start()
     logger.info("API Terminated.")
+
 
 def main(config):
     logging.basicConfig(level=logging.INFO)
     if isinstance(config, str):
         config = ConfigEngine(config)
-
+    Settings(config=config)
     start_api(config)
